Create SulfurBox and Forge version file
The JSON file will be available on my (Vinyl Darkscratch's) web server, updated hourly by the Python script.  mcp_mapping.py grabs a copy of the promotions_slim.json file from Forge, and detects if any Forge versions have updated.  If so, it downloads the MDK of the specified version and obtains the MCP(?) mapping, setting the specified variable.